use CreatePipe on windows

+++ include/cosmo_plugin.hpp
@@ -34,7 +34,7 @@ typedef SSIZE_T ssize_t;
     #pragma warning Unknown dynamic link import/export semantics.
 #endif
 
-extern "C" EXPORT void cosmo_rpc_initialization(int, int);
+extern "C" EXPORT void cosmo_rpc_initialization(long, long);
 
 #endif // __COSMOPOLITAN__
 
@@ -114,7 +114,7 @@ private:
     friend class PluginHost;
 #else
     friend class Plugin;
-    friend void cosmo_rpc_initialization(int, int);
+    friend void cosmo_rpc_initialization(long, long);
 #endif // __COSMOPOLITAN__
 };
 
@@ -150,7 +150,7 @@ public:
 private:
     Plugin();
 
-    friend void cosmo_rpc_initialization(int, int);
+    friend void cosmo_rpc_initialization(long, long);
 };
 
 #endif // __COSMOPOLITAN__
